+++ app/messaging/index.js
@@ -12,7 +12,7 @@ async function start() {
 }
 
 async function stop() {
-  // await plpReceiver.closeConnection();
+ // await plpReceiver.closeConnection();
 }
 
 module.exports = { start, stop };
