+++ app/packing-list/index.js
@@ -2,6 +2,7 @@ const { models, sequelize } = require("../services/database-service");
 const { v4: uuidv4 } = require("uuid");
 
 async function createPackingList(packingListJson, applicationId) {
+
   await sequelize.transaction(async (transaction) => {
     const packingList = packingListMapper(packingListJson, applicationId);
     await models.packingList.create(packingList, {
