+++ app/services/fowlerwelch/parser.js
@@ -1,6 +1,7 @@
 const CombineParser = require("../parser-combine");
 
-function parseFowlerWelch(packingListJson) {
+function parse(packingListJson) {
+  console.log(packingListJson)
   const establishmentNumberRow = 45;
   const establishmentNumber = packingListJson[establishmentNumberRow].M ?? null;
   const packingListContents = packingListJson
@@ -17,4 +18,4 @@ function parseFowlerWelch(packingListJson) {
   return CombineParser.combine(establishmentNumber, packingListContents, true);
 }
 
-module.exports = { parseFowlerWelch };
+module.exports = { parse };
