+++ test/unit/services/fowlerwelch /fowlerwelch-parser2.test.js
@@ -1,0 +1,200 @@
+const fowlerWelchParser2 = require("../../../../app/services/fowlerwelch/parser2");
+
+describe("parseFowlerWelch2", () => {
+  test("parses json", () => {
+    const packingListJson = [
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {
+        A: "Item",
+        B: "Product code",
+        C: "Commodity code",
+        D: "Online Check",
+        E: "Meursing code",
+        F: "Description of goods",
+        G: "Country of Origin",
+        H: "No. of pkgs",
+        I: "Type of pkgs",
+        J: "Total Gross Weight",
+        K: "Total Net Weight",
+        L: "Total Line Value",
+        M: "NIIRMS Dispatch number",
+        N: "Treatment Type (Chilled /Ambient)",
+        O: "NIRMS Lane (R/G)",
+      },
+      {
+        A: "1",
+        B: "6535096",
+        C: "2005800099",
+        D: "",
+        E: "",
+        F: "Sweetcorn Express 8x2",
+        G: "",
+        H: "5",
+        I: "Cases",
+        J: "25.119",
+        K: "20.000",
+        L: "55.40",
+        M: "RMS-GB-000216-002",
+        N: "Chilled",
+        O: "G",
+      },
+      {
+        A: "2",
+        B: "4488002",
+        C: "2005800099",
+        D: "Check - 2005800099",
+        E: "",
+        F: "Sweetcorn Cobettes 24x4",
+        G: "",
+        H: "65",
+        I: "Cases",
+        J: "882.057",
+        K: "780.000",
+        L: "2195.05",
+        M: "RMS-GB-000216-002",
+        N: "Chilled",
+        O: "G",
+      },
+    ];
+
+    const result = fowlerWelchParser2.parse(packingListJson);
+    expect(result.registration_approval_number).toBe(packingListJson[45].M);
+    expect(result.items).toHaveLength(2);
+    expect(result.items[0].description).toBe(packingListJson[45].F);
+    expect(result.items[1].description).toBe(packingListJson[46].F);
+    expect(result.items[0].commodity_code).toBe(packingListJson[45].C);
+    expect(result.items[1].commodity_code).toBe(packingListJson[46].C);
+    expect(result.items[0].number_of_packages).toBe(packingListJson[45].H);
+    expect(result.items[1].number_of_packages).toBe(packingListJson[46].H);
+    expect(result.items[0].total_net_weight_kg).toBe(packingListJson[45].K);
+    expect(result.items[1].total_net_weight_kg).toBe(packingListJson[46].K);
+  });
+
+  test("parses null json", () => {
+    const packingListJson = [
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {},
+      {
+        A: "Item",
+        B: "Product code",
+        C: "Commodity code",
+        D: "Online Check",
+        E: "Meursing code",
+        F: "Description of goods",
+        G: "Country of Origin",
+        H: "No. of pkgs",
+        I: "Type of pkgs",
+        J: "Total Gross Weight",
+        K: "Total Net Weight",
+        L: "Total Line Value",
+        M: "NIIRMS Dispatch number",
+        N: "Treatment Type (Chilled /Ambient)",
+        O: "NIRMS Lane (R/G)",
+      },
+      {
+        A: "1",
+        B: "6535096",
+        D: "",
+        E: "",
+        G: "",
+        I: "Cases",
+        J: "25.119",
+        L: "55.40",
+        O: "G",
+      },
+    ];
+
+    const result = fowlerWelchParser2.parse(packingListJson);
+    expect(result.registration_approval_number).toBeNull();
+    expect(result.items).toHaveLength(1);
+    expect(result.items[0].description).toBeNull();
+    expect(result.items[0].commodity_code).toBeNull();
+    expect(result.items[0].number_of_packages).toBeNull();
+    expect(result.items[0].total_net_weight_kg).toBeNull();
+  });
+});
