+++ test/unit/services/parser-service.test.js
@@ -1753,420 +1753,3 @@ describe("parseBuffaloadLogistics", () => {
     expect(result.items[0].total_net_weight_kg).toBeNull();
   });
 });
-
-describe("matchesFowlerWelch2", () => {
-  test.only("returns generic error for empty json", () => {
-    const packingListJson = {};
-    const filename = "packinglist.xlsx";
-    const result = parserService.matchesFowlerWelch2(packingListJson, filename);
-    expect(result).toBe(MatcherResult.GENERIC_ERROR);
-  });
-  test.only("returns wrong extension for incorrect file extension", () => {
-    const filename = "packinglist.xls";
-    const packingListJson = {};
-    const result = parserService.matchesFowlerWelch2(packingListJson, filename);
-    expect(result).toBe(MatcherResult.WRONG_EXTENSIONS);
-  });
-  test.only("returns wrong establishment number for missing establishment number", () => {
-    const packingListJson = {
-      "Cust Ord - Vitacress": [
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {
-          M: "Incorrect",
-        },
-      ],
-    };
-    const filename = "packinglist.xlsx";
-    const result = parserService.matchesFowlerWelch2(packingListJson, filename);
-    expect(result).toBe(MatcherResult.WRONG_ESTABLISHMENT_NUMBER);
-  });
-  test.only("returns wrong header for incorrect header values", () => {
-    const filename = "packinglist.xlsx";
-    const packingListJson = {
-      "Cust Ord - Vitacress": [
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {
-          A: "NOT",
-          B: "CORRECT",
-          C: "HEADER",
-          D: "Online Check",
-          E: "Meursing Code",
-          F: "Description of goods",
-          G: "Country of Origin",
-          H: "No. of pkgs",
-          I: "Type of pkgs",
-          J: "Total Gross Weight",
-          K: "Total Net Weight",
-          L: "Item Value",
-          M: "NIIRMS Dispatch number",
-          N: "Treatment Type (Chilled /Ambient)",
-          O: "NIRMS Lane (R/G)",
-        },
-        {
-          M: "RMS-GB-000216-002",
-        },
-      ],
-    };
-    const result = parserService.matchesFowlerWelch2(packingListJson, filename);
-    expect(result).toBe(MatcherResult.WRONG_HEADER);
-  });
-  test.only("returns correct for correct headers", () => {
-    const filename = "packinglist.xlsx";
-    const packingListJson = {
-      "Cust Ord - Vitacress": [
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {
-          A: "Item",
-          B: "Product code",
-          C: "Commodity code",
-          D: "Online Check",
-          E: "Meursing code",
-          F: "Description of goods",
-          G: "Country of Origin",
-          H: "No. of pkgs",
-          I: "Type of pkgs",
-          J: "Total Gross Weight",
-          K: "Total Net Weight",
-          L: "Total Line Value",
-          M: "NIIRMS Dispatch number",
-          N: "Treatment Type (Chilled /Ambient)",
-          O: "NIRMS Lane (R/G)",
-        },
-        {
-          M: "RMS-GB-000216-002",
-        },
-      ],
-    };
-    const result = parserService.matchesFowlerWelch2(packingListJson, filename);
-    expect(result).toBe(MatcherResult.CORRECT);
-  });
-
-  describe("parseFowlerWelch2", () => {
-    test.only("parses json", () => {
-      const packingListJson = [
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {
-          A: "Item",
-          B: "Product code",
-          C: "Commodity code",
-          D: "Online Check",
-          E: "Meursing code",
-          F: "Description of goods",
-          G: "Country of Origin",
-          H: "No. of pkgs",
-          I: "Type of pkgs",
-          J: "Total Gross Weight",
-          K: "Total Net Weight",
-          L: "Total Line Value",
-          M: "NIIRMS Dispatch number",
-          N: "Treatment Type (Chilled /Ambient)",
-          O: "NIRMS Lane (R/G)",
-        },
-        {
-          A: "1",
-          B: "6535096",
-          C: "2005800099",
-          D: "",
-          E: "",
-          F: "Sweetcorn Express 8x2",
-          G: "",
-          H: "5",
-          I: "Cases",
-          J: "25.119",
-          K: "20.000",
-          L: "55.40",
-          M: "RMS-GB-000216-002",
-          N: "Chilled",
-          O: "G",
-        },
-        {
-          A: "2",
-          B: "4488002",
-          C: "2005800099",
-          D: "Check - 2005800099",
-          E: "",
-          F: "Sweetcorn Cobettes 24x4",
-          G: "",
-          H: "65",
-          I: "Cases",
-          J: "882.057",
-          K: "780.000",
-          L: "2195.05",
-          M: "RMS-GB-000216-002",
-          N: "Chilled",
-          O: "G",
-        },
-      ];
-
-      const result = parserService.parseFowlerWelch2(packingListJson);
-      expect(result.registration_approval_number).toBe(packingListJson[45].M);
-      expect(result.items).toHaveLength(2);
-      expect(result.items[0].description).toBe(packingListJson[45].F);
-      expect(result.items[1].description).toBe(packingListJson[46].F);
-      expect(result.items[0].commodity_code).toBe(packingListJson[45].C);
-      expect(result.items[1].commodity_code).toBe(packingListJson[46].C);
-      expect(result.items[0].number_of_packages).toBe(packingListJson[45].H);
-      expect(result.items[1].number_of_packages).toBe(packingListJson[46].H);
-      expect(result.items[0].total_net_weight_kg).toBe(packingListJson[45].K);
-      expect(result.items[1].total_net_weight_kg).toBe(packingListJson[46].K);
-    });
-
-    test.only("parses null json", () => {
-      const packingListJson = [
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {},
-        {
-          A: "Item",
-          B: "Product code",
-          C: "Commodity code",
-          D: "Online Check",
-          E: "Meursing code",
-          F: "Description of goods",
-          G: "Country of Origin",
-          H: "No. of pkgs",
-          I: "Type of pkgs",
-          J: "Total Gross Weight",
-          K: "Total Net Weight",
-          L: "Total Line Value",
-          M: "NIIRMS Dispatch number",
-          N: "Treatment Type (Chilled /Ambient)",
-          O: "NIRMS Lane (R/G)",
-        },
-        {
-          A: "1",
-          B: "6535096",
-          D: "",
-          E: "",
-          G: "",
-          I: "Cases",
-          J: "25.119",
-          L: "55.40",
-          O: "G",
-        },
-      ];
-
-      const result = parserService.parseFowlerWelch2(packingListJson);
-      expect(result.registration_approval_number).toBeNull();
-      expect(result.items).toHaveLength(1);
-      expect(result.items[0].description).toBeNull();
-      expect(result.items[0].commodity_code).toBeNull();
-      expect(result.items[0].number_of_packages).toBeNull();
-      expect(result.items[0].total_net_weight_kg).toBeNull();
-    });
-  });
-});
